Add initial cleanup checks - fixes #6

+++ createDevJob.groovy
@@ -3,8 +3,12 @@ if [ ! -d /var/jenkins_home/jobs/devPipeline/ ]; then
     mkdir /var/jenkins_home/jobs/devPipeline/
     mkdir /var/jenkins_home/jobs/devPipeline/workspace/
 fi
-rm -r /var/jenkins_home/jobs/devPipeline/workspace/*
-cp -r ../pipeline-dev/* /var/jenkins_home/jobs/devPipeline/workspace/
+if [ -e /var/jenkins_home/jobs/devPipeline/workspace/* ]; then
+    rm -r /var/jenkins_home/jobs/devPipeline/workspace/*
+fi
+if [[ ! -z "$(ls -A)" ]] ; then
+    cp -r * /var/jenkins_home/jobs/devPipeline/workspace/
+fi
 '''
